+++ backend/src/plans/fiches/fiches.module.ts
@@ -11,17 +11,20 @@ import { CountByService } from './count-by/count-by.service';
 import { ExportService } from './export/export.service';
 import { FicheActionEtapeRouter } from './fiche-action-etape/fiche-action-etape.router';
 import { FicheActionEtapeService } from './fiche-action-etape/fiche-action-etape.service';
-import FicheService from './fiche.service';
+import FicheActionPermissionsService from './fiche-action-permissions.service';
 import FichesActionUpdateService from './fiches-action-update.service';
 import { FichesActionController } from './fiches-action.controller';
 import { FicheActionBudgetService } from '@/backend/plans/fiches/fiche-action-budget/fiche-action-budget.service';
 import { FicheActionBudgetRouter } from '@/backend/plans/fiches/fiche-action-budget/fiche-action-budget.router';
+import FicheActionNoteService from '@/backend/plans/fiches/fiche-action-note/fiche-action-note.service';
+import FicheActionCreateService from '@/backend/plans/fiches/import/fiche-action-create.service';
+import { FicheActionNoteController } from '@/backend/plans/fiches/fiche-action-note/fiche-action-note.controller';
 
 @Module({
   imports: [forwardRef(() => CollectivitesModule)],
   providers: [
     PlanActionsService,
-    FicheService,
+    FicheActionPermissionsService,
     AxeService,
     FicheActionListService,
     CountByService,
@@ -34,9 +37,11 @@ import { FicheActionBudgetRouter } from '@/backend/plans/fiches/fiche-action-bud
     ExportService,
     FicheActionBudgetService,
     FicheActionBudgetRouter,
+    FicheActionNoteService,
+    FicheActionCreateService
   ],
   exports: [
-    FicheService,
+    FicheActionPermissionsService,
     AxeService,
     PlanActionsService,
     FicheActionListService,
@@ -47,7 +52,9 @@ import { FicheActionBudgetRouter } from '@/backend/plans/fiches/fiche-action-bud
     BulkEditRouter,
     FicheActionBudgetService,
     FicheActionBudgetRouter,
+    FicheActionNoteService,
+    FicheActionCreateService
   ],
-  controllers: [FichesActionController, ExportPlanController],
+  controllers: [FichesActionController, ExportPlanController, FicheActionNoteController],
 })
 export class FichesActionModule {}
